Don't install /run/squid (use systemd's RuntimeDirectory instead).
Debian Policy states that /run is normally cleared at boot time, and
therefore packages must not install files/directories under /run.
Init scripts should be taught to dynamically handle /run instead.

This change uses systemd's RuntimeDirectory and RuntimeDirectoryMode
directives when starting the squid service in order to guarantee that
/run/squid/ will be created with the correct permission.  This has the
added benefit of deleting the directory when the service is stopped.