Don't restart squid by hand on postinst script
When installing/upgrading squid, the service is being restarted
manually in the postinst script, which can break installations that
have the squid apparmor enabled because it will try to restart the
service before reloading the apparmor profile.

There is no reason to restart squid manually, since the restart will
be automatically performed later.